requirements.txt
# Core LangChain and LangGraph packages
langchain==0.3.25
langchain-chroma==0.2.4
langchain-cohere==0.4.4
langchain-community==0.3.20
langchain-core==0.3.60
langchain-experimental==0.3.4
langchain-fireworks==0.3.0
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langgraph-supervisor==0.0.25
langmem==0.0.25
langsmith==0.3.18

# LLM Provider packages
langchain-openai==0.3.17
langchain-anthropic==0.3.13
langchain-groq==0.3.4
openai==1.97.0
anthropic>=0.51.0

# Vector Database and Search packages
langchain-chroma>=0.2.4
chromadb>=1.0.12
langchain-tavily>=0.1.6

# Data Science and Visualization
matplotlib>=3.10.0
pandas>=2.0.0
numpy>=1.26.0

# Web UI Framework
streamlit>=1.39.0

# Web and HTTP utilities
requests>=2.32.0
beautifulsoup4>=4.13.0
bs4>=0.0.2

# Data processing and validation
pydantic>=2.10.0
pydantic-settings>=2.8.0

# Database utilities
sqlalchemy>=2.0.39

# Environment and configuration
python-dotenv>=1.0.0

# Jupyter notebook support
ipykernel>=6.29.0
ipython>=9.0.0
jupyter_client>=8.6.0
jupyter_core>=5.7.0

# Core Python utilities
typing_extensions>=4.12.0
pyyaml>=6.0.0
